feat(frontend) AuthServices were implemented, some corrections were fixed in redirection and rendering
