+++ frontend/css/style.css
@@ -9,6 +9,21 @@ body{
     font-family: Arial, Helvetica, sans-serif;
     background-color: #333339;
 }
+
+.init_register{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+}
+.init_login{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+}
 .feature_container{
     display: grid;
     place-items: center;
