+++ frontend/index.html
@@ -11,9 +11,25 @@
     <title>Gym </title>
 </head>
 <body>
-    
-    <div id="main_content"></div>
-
-    <script type="module" src="/app/app.js"></script>
+    <div class="init_login">
+        <div class="container_btn">
+            <button data-route="/newUser" class="btn_add_user">Añadir Usuario</button>
+        </div>
+        <div class="feature_container">
+            <form class="login_container">
+                <h3>Bienvenido</h3>
+                <div class="input_container">
+                    <label  class="label_user">Email/Usuario</label>
+                    <input placeholder="Ej. pedro@gmail" type="text" id="user" class="field_user" />
+                    <label class="label_user">Contraseña</label>
+                    <input placeholder="************" type="password" id="pass" class="field_user"/>
+                </div>
+                <button class="btn submit_button">Iniciar Sesión</button>
+                <a href="">Olvidaste tu contraseña?</a>
+            </form>
+        </div>
+    </div>
+    <script type="module" src="/controllers/auth/login.js"></script>
 </body>
 </html>
+
