+++ frontend/services/auth.services.js
@@ -1,0 +1,22 @@
+const API_URL="http://localhost:8500";
+
+export const loginService=async(email, password)=>{
+    try {
+    const response = await fetch(`${API_URL}/api/auth/login`, {
+                method: 'POST',
+                headers:{
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify({email, password}),
+                credentials: "include"
+            })
+
+    if(!response.ok) throw new Error('Invalid email or password');
+    const userData=response.json();
+    return userData;
+
+  } catch (error) {
+    window.location.href = '/login'
+    return null
+  }
+}
