+++ frontend/views/auth/registUser.html
@@ -1,43 +1,57 @@
-
-    <div class="container_btn">
-        <button data-route="/login" class="btn_init_log">Iniciar sesion</button>
-    </div>
-    <div class="feature_container">
-        <form class="login_container">
-            <h3>Registrar Usuario</h3>
-            <div class="input_container">
-              <label class="label_user">Rol</label>
-              <select class="field_user f_role" id="role">
-                <option value="role0" class="role_option">Elegir Rol</option>
-                <option value="0"  class="role_option">Administrador</option>
-                <option value="1"  class="role_option">Empleado</option>
-              </select>
-              <p class="error" id="role_error"></p>  
-            </div>
-            <div class="input_container">
-              <label for="email" class="label_user">Email</label>
-              <select name="selected_email" class="field_user f_email" id="email" >
-                <option value="email0" disabled selected>Elegir Correo</option> 
-              </select>
-            </div>
-            <p class="error" id="email_error"></p>  
-            <div class="input_container">
-              <label class="label_user">Contraseña</label>
-              <input placeholder="************" type="password" id="password" class="field_user"/>
-            </div>
-            <p class="error" id="password_error"></p>  
-            <!--Button group-->
-            <div class="dialog-container">
-              <dialog id="alert-dialog">
-              <h3>Confirmar Registro</h3>
-              <div class="button_container_modal">
-                <button id="close_modal" class="btn_submit_modal cancel">Cerrar</button>
-                <button id="modal_submit" class="btn_submit_modal confirm">Confirmar</button>
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <link rel="stylesheet" href="../../css/style.css">
+  <title>Document</title>
+</head>
+<body>
+    <div class="init_register">
+      <div class="container_btn">
+          <button class="btn_init_log">Iniciar sesion</button>
+      </div>
+      <div class="feature_container">
+          <form class="login_container">
+              <h3>Registrar Usuario</h3>
+              <div class="input_container">
+                <label class="label_user">Rol</label>
+                <select class="field_user f_role" id="role">
+                  <option value="role0" class="role_option">Elegir Rol</option>
+                  <option value="0"  class="role_option">Administrador</option>
+                  <option value="1"  class="role_option">Empleado</option>
+                </select>
+                <p class="error" id="role_error"></p>  
               </div>
-            </dialog>
-            </div>
-            <div class="toast_container"></div>
-            <button class="btn submit_button">Registrar Usuario</button>
-            <a href="" >Iniciar Sesión</a>
-          </form>
-    </div>
+              <div class="input_container">
+                <label for="email" class="label_user">Email</label>
+                <select name="selected_email" class="field_user f_email" id="email" >
+                  <option value="email0" disabled selected>Elegir Correo</option> 
+                </select>
+              </div>
+              <p class="error" id="email_error"></p>  
+              <div class="input_container">
+                <label class="label_user">Contraseña</label>
+                <input placeholder="************" type="password" id="password" class="field_user"/>
+              </div>
+              <p class="error" id="password_error"></p>  
+              <!--Button group-->
+              <div class="dialog-container">
+                <dialog id="alert-dialog">
+                <h3>Confirmar Registro</h3>
+                <div class="button_container_modal">
+                  <button id="close_modal" class="btn_submit_modal cancel">Cerrar</button>
+                  <button id="modal_submit" class="btn_submit_modal confirm">Confirmar</button>
+                </div>
+              </dialog>
+              </div>
+              <div class="toast_container"></div>
+              <button class="btn submit_button">Registrar Usuario</button>
+              <a href="../../index.html" >Iniciar Sesión</a>
+            </form>
+      </div>
+    </div>   
+    
+    <script type="module" src="../../controllers/auth/registUser.js"></script>
+</body>
+</html>
